requirements.txt
fastapi>=0.104.0
typer>=0.9.0
networkx>=3.1
pydantic>=2.0
sqlmodel>=0.0.14
jinja2>=3.1
graphviz>=0.20
pyyaml>=6.0
python-hcl2>=0.4.0
uvicorn>=0.24.0
